% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{class_proportions}
\alias{class_proportions}
\title{Computes class proportions}
\usage{
class_proportions(x, possibleValues)
}
\arguments{
\item{x}{A numeric vector.}

\item{possibleValues}{A vector of all possible values x can contain}
}
\description{
Computes class proportions
}
\keyword{internal}
