% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann.R
\name{dann.formula}
\alias{dann.formula}
\title{Discriminant Adaptive Nearest Neighbor Classification}
\usage{
\method{dann}{formula}(
  formula,
  data,
  k = 5,
  neighborhood_size = max(floor(nrow(data)/5), 50),
  epsilon = 1,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the outcome and predictors. For example, Y ~ X1 + X2.}

\item{data}{A data frame containing the variables in \code{formula} or in the recipe.}

\item{k}{The number of nearest neighbors used to classify a point. Identical to k in standard k nearest neighbors.}

\item{neighborhood_size}{The number of nearest neighbors used to estimate the between and within class covariance matrices that shape the neighborhood.}

\item{epsilon}{Softening parameter. Scales the identity matrix added to the between class covariance, which keeps the neighborhood from collapsing onto the class boundary. 1 matches the publication.}

\item{...}{Additional parameters passed to methods.}
}
\value{
An S3 class of type dann.
}
\description{
Discriminant Adaptive Nearest Neighbor Classification
}
\details{
This is an implementation of Hastie and Tibshirani's
\href{https://web.stanford.edu/~hastie/Papers/dann_IEEE.pdf}{Discriminant Adaptive Nearest Neighbor Classification}.
}
\examples{
library(dann)
library(mlbench)
library(magrittr)
library(dplyr)

set.seed(1)
train <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(train) <- c("X1", "X2", "Y")

dann(Y ~ X1 + X2, train)
}
