% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann2.R
\name{dann2}
\alias{dann2}
\title{dann2}
\usage{
dann2(x, testx = matrix(double(p), nrow = 1), y, k = 5,
  kmetric = length(y)/2, epsilon = 1, rate = 0.5, fullw = FALSE,
  scalar = FALSE, iter = 1, covmin = 1e-04, cv = FALSE)
}
\arguments{
\item{x}{xmat}

\item{testx}{test mat}

\item{y}{labels}

\item{k}{nearest neighbors}

\item{kmetric}{kmet}

\item{epsilon}{epsilon}

\item{rate}{rate}

\item{fullw}{fullw}

\item{scalar}{scalar}

\item{iter}{iter}

\item{covmin}{covmin}

\item{cv}{cv}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
dann2
}

