% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{DANN_distance}
\alias{DANN_distance}
\title{Computes the distance between x0 and x1 using the DANN metric}
\usage{
DANN_distance(x0, x1, sigma)
}
\arguments{
\item{x1}{A numeric matrix with training predictors as columns.}

\item{sigma}{A numeric matrix defined in Hastie's DANN publication.}

\item{x2}{A numeric matrix with training predictors as columns.}
}
\description{
Computes the distance between x0 and x1 using the DANN metric
}
\keyword{internal}
