% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dannsub.R
\name{dannsub}
\alias{dannsub}
\title{dannsub}
\usage{
dannsub(x, y, km = max(50, 0.2 * n), k = 5, epsilon = 1, fullw = FALSE,
  scalex = TRUE, scalar = TRUE, dims = seq(p), iter.sub = 1, iter = 1,
  method = "knn", xnew, ynew)
}
\arguments{
\item{x}{covariate matrix}

\item{y}{labels}

\item{km}{max subspace}

\item{k}{number of clusters}

\item{epsilon}{error}

\item{fullw}{Boolean}

\item{scalex}{Boolean, default true}

\item{scalar}{Boolean,}

\item{iter.sub}{Max. number of subspace iterations}

\item{iter}{Max number of iterations}

\item{method}{knn}

\item{xnew}{n}

\item{ynew}{}
}
\value{
an object of class dannsub
}
\description{
dannsub
}
\examples{
\dontrun {
dannsub(km = max(50, 0.2 * n), k = 5, epsilon = 1, fullw = FALSE, scalex = TRUE, 
   scalar = TRUE, dims = seq(p), iter.sub = 1, iter = 1, method = "knn", xnew, ynew)
}
}

