% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dannsubauto.R
\name{dannsubauto}
\alias{dannsubauto}
\title{dannsubauto}
\usage{
dannsubauto(x, testx = matrix(double(p), nrow = 1), y, k = 5,
  kmetric = max(0.2 * n, 50), epsilon = 1, plus = 1, trace = FALSE, ...)
}
\arguments{
\item{x}{mat}

\item{testx}{test mat}

\item{y}{labels}

\item{k}{k}

\item{kmetric}{kmet}

\item{epsilon}{epsilon}

\item{plus}{1}

\item{trace}{Boolean}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
dannsubauto
}

