% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon.list.R
\docType{data}
\name{epsilon.list}
\alias{epsilon.list}
\title{Epsilon values for DANN}
\format{
Numeric vector of 10 epsilon values: 1, 0.001, 0.01, 0.1, 0.5, 1, 2, 3, 4, 5
}
\usage{
epsilon.list
}
\description{
Epsilon values for DANN
}
\keyword{data}
