% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.test.R
\docType{data}
\name{glass.test}
\alias{glass.test}
\title{Test data for glass}
\format{List of two elements: x - 96 x 9 training data matrix, y - vector of labels}
\usage{
glass.test
}
\description{
Test data for glass
}
\keyword{data}

