% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glass.train.R
\docType{data}
\name{glass.train}
\alias{glass.train}
\title{Training data for glass}
\format{List of two elements: x - 89 x 9 training data matrix, y - vector of labels}
\usage{
glass.train
}
\description{
Training data for glass
}
\keyword{data}

