% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{knn}
\usage{
knn(train, test, cl, k = 1)
}
\arguments{
\item{train}{matrix}

\item{test}{test}

\item{cl}{number of clusters}

\item{k}{number of clusters}
}
\value{
result
}
\description{
K-nearest Neighbors
}
\examples{
\dontrun{ 
knn()
}
}

