% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knncv.R
\name{knncv}
\alias{knncv}
\title{knncv}
\usage{
\dontrun{
knncv(x, y, k=5)
}
}
\arguments{
\item{x}{matrix}

\item{y}{}

\item{k}{number of clusters}
}
\value{
list of standardized matrices along with scaling information (centers)
}
\description{
Cross-validated k-means
}

