% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann.R
\name{mode}
\alias{mode}
\title{Computes mode.
Code found at \href{https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{Stack Overflow}}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Should na be removed?}
}
\description{
Computes mode.
Code found at \href{https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{Stack Overflow}
}
\keyword{internal}
