% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydann.R
\name{mydann}
\alias{mydann}
\title{mydann}
\usage{
mydann(train.data, test.data, p = dim(x)[2], kmetric = max(50, 0.2 * n),
  k = 5, epsilon.list = c(1, 0.5, 2, 5), iter.list = 1, ...)
}
\arguments{
\item{train.data}{training data. All the features in matrix x. and class labels in y.}

\item{test.data}{test data. Same format as the test data}

\item{p}{dimensions of x}

\item{kmetric}{k met}

\item{k}{Number of nearest neighbors}

\item{epsilon.list}{epsilon}

\item{iter.list}{iteration}
}
\value{
results
}
\description{
mydann
}
\examples{
\dontrun{ 
mydann()
}
}

