% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndist2.R
\name{nndist2}
\alias{nndist2}
\title{nndist2}
\usage{
# dann()
}
\arguments{
\item{x}{matrix}

\item{y}{y}

\item{x0}{colMeans of x}

\item{ktarget}{number of clusters}

\item{rate}{rate of learning}

\item{epsilon}{epsilon}

\item{fullw}{fullw}

\item{scalar}{boolean}

\item{iter}{number of iterations}

\item{cv}{cross-validate}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
Main function
}

