% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnsubmeans.R
\name{nnsubmeans}
\alias{nnsubmeans}
\title{nnsubmeans}
\usage{
# dann()
}
\arguments{
\item{x}{matrix}

\item{y}{y}

\item{k}{number of clusters}

\item{epsilon}{epsilon}

\item{fullw}{fullw}

\item{iter}{number of iterations}

\item{cv}{cross-validate}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
Main function
}

