% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnsubspace.R
\name{nnsubspace}
\alias{nnsubspace}
\title{nnsubspace}
\usage{
nnsubspace(x, y, k = length(y)/2, epsilon = 1, fullw = FALSE, iter = 1,
  cv = 0)
}
\arguments{
\item{x}{mat}

\item{y}{labels}

\item{k}{clusters}

\item{epsilon}{epsilon}

\item{fullw}{fullw}

\item{iter}{iter}

\item{cv}{cv}
}
\value{
subspace
}
\description{
nnsubspace
}
\examples{
\dontrun{
dannsub()
}
}

