% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dann.R
\name{predict.sub_dann}
\alias{predict.sub_dann}
\title{Discriminant Adaptive Nearest Neighbor With Subspace Reduction}
\usage{
\method{predict}{sub_dann}(object, new_data, type = "class", ...)
}
\arguments{
\item{object}{A fitted model of class sub_dann.}

\item{new_data}{A data frame of predictors to score.}

\item{type}{Type of prediction. One of "class" or "prob".}

\item{...}{Not used.}
}
\value{
A data frame of predicted classes or class probabilities. Adheres to tidymodels standards.
}
\description{
Discriminant Adaptive Nearest Neighbor With Subspace Reduction
}
\details{
An implementation of Hastie and Tibshirani's sub-dann in section 4.1 of
\href{https://web.stanford.edu/~hastie/Papers/dann_IEEE.pdf}{Discriminant Adaptive Nearest Neighbor Classification}.

dann's performance suffers when unrelated variables are included in the model. sub_dann first
projects the predictors onto a lower dimensional subspace found by \code{\link[fpc:ncoord]{fpc::ncoord()}} and then fits
dann on that subspace. Simulations show sub_dann generally performs better in this scenario.
}
\examples{
library(dann)
library(mlbench)
library(magrittr)
library(dplyr)

set.seed(1)
train <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(train) <- c("X1", "X2", "Y")

test <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(test) <- c("X1", "X2", "Y")

model <- sub_dann(Y ~ X1 + X2, train)
predict(model, test, "class")

predict(model, test, "prob")
}
