% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sub_dann}
\alias{print.sub_dann}
\title{Print sub_dann model}
\usage{
\method{print}{sub_dann}(x, ...)
}
\arguments{
\item{x}{A sub_dann model.}

\item{...}{Not used.}
}
\value{
The model, invisibly. Called for the side effect of printing.
}
\description{
Print sub_dann model
}
\examples{
library(dann)
library(mlbench)
library(magrittr)
library(dplyr)

set.seed(1)
train <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(train) <- c("X1", "X2", "Y")

model <- sub_dann(Y ~ X1 + X2, train)
print(model)
}
