% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vowel.R
\docType{data}
\name{Vowel}
\alias{Vowel}
\title{Vowel data}
\format{Data frame with 990 rows and 11 variables
Column Names: "V1", "V2", "V3", "V4", "V5", "V6", "V7", "V8", "V9", "V10", "Class"
V1 and Class are factors while the rest are numeric}
\usage{
Vowel
}
\description{
Vowel data
}
\keyword{data}

