% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann-package.R
\docType{package}
\name{dann-package}
\alias{dann-package}
\title{dann: Discriminant Adaptive Nearest Neighbors}
\description{
Discriminant Adaptive Nearest Neighbors
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soodoku/dann}
  \item Report bugs at \url{https://github.com/soodoku/dann/issues}
}

}
\author{
Trevor Hastie
}
