% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann.R
\name{dann}
\alias{dann}
\title{Discriminant Adaptive Nearest Neighbor Classification}
\usage{
dann(xTrain, yTrain, xTest, k = 5,
  neighborhood_size = min(floor(nrow(xTrain)/5), 50), epsilon = 1,
  probability = FALSE)
}
\arguments{
\item{xTrain}{Train features. Something easily converted to a numeric matrix.}

\item{yTrain}{Train classes. Something easily converted to a numeric matrix.}

\item{xTest}{Test features. Something easily converted to a numeric matrix.}

\item{k}{The number of data points used for final classificastion.}

\item{neighborhood_size}{The number of data points used to calcualate between and within class covariance.}

\item{epsilon}{Diaginal elemnts of a diagional matrix. 1 is the identity matirx.}

\item{probability}{Should probabilities instead of classes be returned?}
}
\value{
A numeric matrix containing class predictions.
}
\description{
Discriminant Adaptive Nearest Neighbor Classification
}
\details{
This is an implementation of Hastie and Tibshirani's
\href{https://web.stanford.edu/~hastie/Papers/dann_IEEE.pdf}{Discriminant Adaptive Nearest
Neighbor Classificastion publication.}.
The code is a port of Christopher Jenness's
python \href{https://github.com/christopherjenness/ML-lib}{implementation.}
}
