% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dann.R
\name{dann}
\alias{dann}
\title{dann}
\usage{
# dann()
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), No. of fields remaining (given the money), and when the application credits expire.
}
\description{
dann
}

