% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann2.R
\name{dann2}
\alias{dann2}
\title{dann2}
\usage{
# dann2()
}
\arguments{
\item{k}{}

\item{kmetric}{}

\item{epsilon}{}

\item{rate}{}

\item{cv}{}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
dann2
}

