% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dannsub.R
\name{dannsub}
\alias{dannsub}
\title{dannsub}
\usage{
# dannsub()
}
\arguments{
\item{x}{}

\item{y}{}

\item{km}{}

\item{k}{}

\item{epsilon}{}

\item{fullw}{}

\item{scalex}{}

\item{scalar}{}

\item{dims}{}

\item{iter.sub}{}

\item{iter}{}

\item{method}{}

\item{xnew}{}

\item{ynew}{}

\item{iter}{}
}
\value{
an object of class dannsub
}
\description{
dannsub
}

