% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dannsubauto.R
\name{dannsubauto}
\alias{dannsubauto}
\title{dannsubauto}
\usage{
# dannsub()
}
\arguments{
\item{x}{}

\item{y}{}

\item{k}{}

\item{epsilon}{}

\item{km}{}

\item{fullw}{}

\item{scalex}{}

\item{scalar}{}

\item{iter}{}

\item{dims}{}

\item{iter.sub}{}

\item{iter}{}

\item{method}{}

\item{xnew}{}

\item{ynew}{}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money),
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
dannsubauto
}

