% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{MODE}
\alias{MODE}
\title{Computes mode.
Code found at \href{https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{Stack Overflow}}
\usage{
MODE(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Computes mode.
Code found at \href{https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{Stack Overflow}
}
\keyword{internal}
