% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydann.R
\name{mydann}
\alias{mydann}
\title{min2d}
\usage{
mydann(train.data, test.data, p = dim(x)[2], kmetric = max(50, 0.2 * n),
  k = 5, epsilon.list = c(1, 0.5, 2, 5), iter.list = 1, ...)
}
\arguments{
\item{mat}{matrix}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
min2d
}
\examples{
mydann()
}

