% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mydann.R
\name{mydann}
\alias{mydann}
\title{min2d}
\usage{
min2d()
}
\arguments{
\item{mat}{matrix}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money),
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
min2d
}

