% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndist.R
\name{nndist}
\alias{nndist}
\title{nndist}
\usage{
nndist(x, y, x0 = apply(x, 2, mean), k = length(y)/2, epsilon = 1,
  fullw = FALSE, scalar = FALSE, iter = 1, cv = 0)
}
\arguments{
\item{x}{matrix}

\item{y}{y}

\item{x0}{colMeans of x}

\item{k}{number of clusters}

\item{epsilon}{epsilon}

\item{fullw}{fullw}

\item{scalar}{boolean}

\item{iter}{number of iterations}

\item{cv}{cross-validate}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
Main function
}
\examples{
\dontrun{
nndist()
}
}

