% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nnsubspace.R
\name{nnsubspace}
\alias{nnsubspace}
\title{nnsubspace}
\usage{
# dannsub()
}
\arguments{
\item{x}{}

\item{y}{}

\item{k}{}

\item{epsilon}{}

\item{fullw}{}

\item{iter}{}

\item{cv}{}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money),
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
nnsubspace
}

