% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnsubspace.R
\name{nnsubspace}
\alias{nnsubspace}
\title{nnsubspace}
\usage{
# dannsub()
}
\arguments{
\item{k}{}

\item{epsilon}{}

\item{cv}{}
}
\value{
A list with items including Name of the Application, No. of pages remaining (given the money), 
No. of fields remaining (given the money), and when the application credits expire.
}
\description{
nnsubspace
}

