% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dann.R
\name{predict.dann}
\alias{predict.dann}
\title{Discriminant Adaptive Nearest Neighbor Classification}
\usage{
\method{predict}{dann}(object, new_data, type = "class", ...)
}
\arguments{
\item{object}{A fitted model of class dann.}

\item{new_data}{A data frame of predictors to score.}

\item{type}{Type of prediction. One of "class" or "prob".}

\item{...}{Not used.}
}
\value{
A data frame of predicted classes or class probabilities. Adheres to tidymodels standards.
}
\description{
Discriminant Adaptive Nearest Neighbor Classification
}
\details{
This is an implementation of Hastie and Tibshirani's
\href{https://web.stanford.edu/~hastie/Papers/dann_IEEE.pdf}{Discriminant Adaptive Nearest Neighbor Classification}.
}
\examples{
library(dann)
library(mlbench)
library(magrittr)
library(dplyr)

set.seed(1)
train <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(train) <- c("X1", "X2", "Y")

test <- mlbench.circle(300, 2) \%>\%
  tibble::as_tibble()
colnames(test) <- c("X1", "X2", "Y")

model <- dann(Y ~ X1 + X2, train)
predict(model, test, "class")

predict(model, test, "prob")
}
