% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.R
\name{stand}
\alias{stand}
\title{stand}
\usage{
stand(x = NA, xx = NA)
}
\arguments{
\item{x}{matrix}

\item{xx}{matrix with same number of columns as x}
}
\value{
list of standardized matrices along with scaling information (centers)
}
\description{
Standardize each column of the matrix
}
\examples{
stand(matrix(1:6, ncol=3), matrix(1:6, ncol=3))
}

